requirements.txt
apache-airflow
pandas
os
gcloud
pyarrow
db_dtypes